Fix stepchild age label
